tests

+++ tests/BaseTest.php
@@ -8,11 +8,20 @@ abstract class BaseTest extends \PHPUnit\Framework\TestCase
 	{
         $dir = __DIR__.'/../var';
 
-        $di = new \RecursiveDirectoryIterator($dir, \FilesystemIterator::SKIP_DOTS);
-		$ri = new \RecursiveIteratorIterator($di, \RecursiveIteratorIterator::CHILD_FIRST);
-		foreach ( $ri as $file ) {
-		    $file->isDir() ?  rmdir($file) : unlink($file);
+        if (file_exists($dir)) {
+	        $di = new \RecursiveDirectoryIterator($dir, \FilesystemIterator::SKIP_DOTS);
+			$ri = new \RecursiveIteratorIterator($di, \RecursiveIteratorIterator::CHILD_FIRST);
+			foreach ( $ri as $file ) {
+			    $file->isDir() ?  rmdir($file) : unlink($file);
+			}
+			rmdir($dir);
 		}
 
+		mkdir($dir, 0777, true);
+	}
+
+	public function getDir()
+	{
+		return __DIR__."/../var/cache";
 	}
 }
